Add actix framework generator option
