+++ generator/src/templates/actix_cargo.toml
@@ -1,23 +1,22 @@
 [package]
 edition = '2021'
-name = 'seaography-sqlite-example'
-version = '0.2.0'
+name = '<seaography-package-name>'
+version = '0.1.0'
 
 [dependencies]
-poem = { version = "1.3.29" }
+actix-web = { version = "4.0.1", default-features = false, features = ["macros"] }
 async-graphql = { version = "4.0.14", features = ["decimal", "chrono", "dataloader"] }
-async-graphql-poem = { version = "4.0.14" }
+async-graphql-actix-web = { version = "4.0.14" }
 async-trait = { version = "0.1.53" }
 dotenv = "0.15.0"
-sea-orm = { version = "^0.9", features = ["sqlx-sqlite", "runtime-async-std-native-tls"] }
+sea-orm = { version = "^0.9", features = ["<seaography-sql-library>", "runtime-async-std-native-tls"] }
 tokio = { version = "1.17.0", features = ["macros", "rt-multi-thread"] }
 tracing = { version = "0.1.34" }
 tracing-subscriber = { version = "0.3.11" }
 lazy_static = { version = "1.4.0" }
 
 [dependencies.seaography]
-path = "../../" # remove this line in your own project
-version = "^0.2" # seaography version
+version = "<seaography-version>" # seaography version
 features = ["with-decimal", "with-chrono"]
 
 [dev-dependencies]
