+++ generator/src/templates/mod.rs
@@ -1,0 +1,2 @@
+pub mod actix;
+pub mod poem;
