+++ generator/src/templates/poem_cargo.toml
@@ -4,11 +4,11 @@ name = '<seaography-package-name>'
 version = '0.2.0'
 
 [dependencies]
+poem = { version = "1.3.29" }
 async-graphql = { version = "4.0.14", features = ["decimal", "chrono", "dataloader"] }
 async-graphql-poem = { version = "4.0.14" }
 async-trait = { version = "0.1.53" }
 dotenv = "0.15.0"
-poem = { version = "1.3.29" }
 sea-orm = { version = "^0.9", features = ["<seaography-sql-library>", "runtime-async-std-native-tls"] }
 tokio = { version = "1.17.0", features = ["macros", "rt-multi-thread"] }
 tracing = { version = "0.1.34" }
