+++ generator/src/writer.rs
@@ -1,7 +1,7 @@
 use proc_macro2::TokenStream;
 use quote::quote;
 
-use crate::util::add_line_break;
+use crate::{util::add_line_break, WebFrameworkEnum};
 
 pub fn generate_query_root(
     entities_hashmap: &crate::sea_orm_codegen::EntityHashMap,
@@ -44,6 +44,7 @@ pub fn write_cargo_toml<P: AsRef<std::path::Path>>(
     path: &P,
     crate_name: &str,
     sql_library: &str,
+    framework: WebFrameworkEnum,
 ) -> std::io::Result<()> {
     let file_path = path.as_ref().join("Cargo.toml");
 
@@ -53,7 +54,12 @@ pub fn write_cargo_toml<P: AsRef<std::path::Path>>(
         env!("CARGO_PKG_VERSION_MINOR")
     );
 
-    let content = include_str!("_Cargo.toml")
+    let content = match framework {
+        WebFrameworkEnum::Actix => include_str!("./templates/actix_cargo.toml"),
+        WebFrameworkEnum::Poem => include_str!("./templates/poem_cargo.toml"),
+    };
+
+    let content = content
         .replace("<seaography-package-name>", crate_name)
         .replace("<seaography-sql-library>", sql_library)
         .replace("<seaography-version>", &ver);
@@ -85,96 +91,6 @@ pub fn write_lib<P: AsRef<std::path::Path>>(path: &P) -> std::io::Result<()> {
     let tokens = generate_lib();
 
     let file_name = path.as_ref().join("lib.rs");
-
-    std::fs::write(file_name, add_line_break(tokens))?;
-
-    Ok(())
-}
-
-///
-/// Used to generate project/src/main.rs file content
-///
-pub fn generate_main(crate_name: &str) -> TokenStream {
-    let crate_name_token: TokenStream = crate_name.replace('-', "_").parse().unwrap();
-
-    quote! {
-        use async_graphql::{
-            dataloader::DataLoader,
-            http::{playground_source, GraphQLPlaygroundConfig},
-            EmptyMutation, EmptySubscription, Schema,
-        };
-        use async_graphql_poem::GraphQL;
-        use dotenv::dotenv;
-        use lazy_static::lazy_static;
-        use poem::{get, handler, listener::TcpListener, web::Html, IntoResponse, Route, Server};
-        use sea_orm::Database;
-        use #crate_name_token::*;
-        use std::env;
-
-        lazy_static! {
-            static ref URL: String = env::var("URL").unwrap_or("0.0.0.0:8000".into());
-            static ref ENDPOINT: String = env::var("ENDPOINT").unwrap_or("/".into());
-            static ref DATABASE_URL: String =
-                env::var("DATABASE_URL").expect("DATABASE_URL environment variable not set");
-            static ref DEPTH_LIMIT: Option<usize> = env::var("DEPTH_LIMIT").map_or(None, |data| Some(
-                data.parse().expect("DEPTH_LIMIT is not a number")
-            ));
-            static ref COMPLEXITY_LIMIT: Option<usize> = env::var("COMPLEXITY_LIMIT")
-                .map_or(None, |data| {
-                    Some(data.parse().expect("COMPLEXITY_LIMIT is not a number"))
-                });
-        }
-
-        #[handler]
-        async fn graphql_playground() -> impl IntoResponse {
-            Html(playground_source(GraphQLPlaygroundConfig::new(&*ENDPOINT)))
-        }
-
-        #[tokio::main]
-        async fn main() {
-            dotenv().ok();
-            tracing_subscriber::fmt()
-                .with_max_level(tracing::Level::INFO)
-                .with_test_writer()
-                .init();
-
-            let database = Database::connect(&*DATABASE_URL)
-                .await
-                .expect("Fail to initialize database connection");
-            let orm_dataloader: DataLoader<OrmDataloader> = DataLoader::new(
-                OrmDataloader {
-                    db: database.clone(),
-                },
-                tokio::spawn,
-            );
-            let mut schema = Schema::build(QueryRoot, EmptyMutation, EmptySubscription)
-                .data(database)
-                .data(orm_dataloader);
-            if let Some(depth) = *DEPTH_LIMIT {
-                schema = schema.limit_depth(depth);
-            }
-            if let Some(complexity) = *COMPLEXITY_LIMIT {
-                schema = schema.limit_complexity(complexity);
-            }
-            let schema = schema.finish();
-            let app = Route::new().at(
-                &*ENDPOINT,
-                get(graphql_playground).post(GraphQL::new(schema)),
-            );
-
-            println!("Visit GraphQL Playground at http://{}", *URL);
-            Server::new(TcpListener::bind(&*URL))
-                .run(app)
-                .await
-                .expect("Fail to start web server");
-        }
-    }
-}
-
-pub fn write_main<P: AsRef<std::path::Path>>(path: &P, crate_name: &str) -> std::io::Result<()> {
-    let tokens = generate_main(crate_name);
-
-    let file_name = path.as_ref().join("main.rs");
 
     std::fs::write(file_name, add_line_break(tokens))?;
